Destroy the Wrapping Modal as well, but don't destroy Semantic
The wrapper was still staying active as a reactive view because it was
not getting removed with Blaze.remove.  Using Blaze.remove gets rid of
the need to call the remove() methods on the modal.

Also, changed to use onHidden instead of onHide.  onHidden runs AFTER
the modal has been hidden (as decided by Semantic, who knows best).

No longer removing the page dimmer because this destroys any other
modals that may have been created outside this library.  It's up to
Semantic-UI to decide what to do with that page dimmer, as should be
left to them.  It doesn't hurt if it stays there, as it hides it
automatically and it has no bindings from this package.