Removed Hardcoded Modal HTML IDs.
This will, in theory, allow you to turn on allowMultiple and have
multiple modals stacked or in place at the same time through this
library.  I haven't, however, changed it to allow options passed
through to Semantic-UI's modal() method.  But if you change it by
hand it does, in fact, work. :)